OPTIM: freq-ctr: don't take the date lock for most updates
It's amazing that the value was still incremented under the date lock,
let's first use an atomic increment for the counter and move it out of
the date lock to reduce contention. These are just counters, we don't
need to take locks if we're not rotating, atomic ops are enough. This
patch does this, and leaves the lock for when the period is over. It's
important to note that some values might be added just before or just
after a rotation but this is not a problem since we don't care if a
value is counted in the previous or next period when it's exactly on
the edge. Great care was taken to ensure that the current counter is
always atomically updated.

Other minor cleanups were performed, such as avoiding to reload the
value from memory after a CAS, or using &~1 instead of two shifts to
remove the lowest bit.

(cherry picked from commit 0d6c75d7495091a615603152993378a2274a76b5)
[ad: this commit was taken to ease backport of the following fix on
freq-ct,
minor adjustments were made: replace _HA_ATOMIC* macros by HA_ATOMIC*
and remove unneeded barriers in context]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>