BUILD: mworker: silence two printf format warnings around getpid()
getpid() is documented as returning a pit pid_t result, not
necessarily an int. This causes a build warning on Solaris 10
because of '%d' or '%u' are used in the format passed to snprintf().
Let's just cast the result as an int (respectively unsigned int).

This can be backported to 2.0 and possibly older versions though
it really has no impact.

(cherry picked from commit 76a80c710cbd8ca3d7dcd191d26cc7a450dee22e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>