BUG/MEDIUM: mworker/cli: command pipelining doesn't work anymore
Since commit 829bd471 ("MEDIUM: stream: rearrange the events to remove
the loop"), the pipelining in the master CLI does not work anymore.

Indeed when doing:

  echo "@1 show info; @2 show info; @3 show info" | socat /tmp/haproxy.master -

the CLI will only show the response of the first command.

When debugging we can observe that the command is sent, but the client
closes the connection before receiving the response.

The problem is that the flag CF_READ_NULL is not cleared when we
reiniate the flags of the response and we rely on this flag to close.

Must be backported in 2.0

(cherry picked from commit 19338011368b787b18598aac00bd639657f22bfa)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>