BUG/MINOR: mux-h1: Don't process input or ouput if an error occurred
It is useless to proceed if an error already occurred. Instead, it is better to
wait it will be catched by the stream or the connection, depending on which is
the first one to detect it.

This patch must be backported to 2.0.

(cherry picked from commit 0e54d547f11a7b51fa27c334f666c312d0c2625a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>