MINOR: stream-int: Factorize processing done after sending data in si_cs_send()
In the function si_cs_send(), what is done when an error occurred on the
connection or the conn_stream or when some successfully data was send via a pipe
or the channel's buffer may be factorized at the function. It slightly simplify
the function.

This patch must be backported to 2.0 and 1.9 because a bugfix depends on it.

(cherry picked from commit 86162db15c439479d154b352d40d2477f39f4fcf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>