BUG/MEDIUM: checks: Don't attempt to read if we destroyed the connection.
In event_srv_chk_io(), only call __event_srv_chk_r() if we did not subscribe
for reading, and if wake_srv_chk() didn't return -1, as it would mean it
just destroyed the connection and the conn_stream, and attempting to use
those to recv data would lead to a crash.

This should be backported to 1.9 and 2.0.

(cherry picked from commit bc89ad8d94246eca064b87e76c1897f6a950f6fb)
Signed-off-by: Willy Tarreau <w@1wt.eu>