BUG/MINOR: cache/htx: Make maxage calculation HTX aware
The function http_calc_maxage() was not updated to be HTX aware. So the header
"Cache-Control" on the response was never parsed to find "max-age" or "s-maxage"
values.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit 5f2c49f5eedb55fd757c2e4812df840bcf2dd9e9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>