BUG/MEDIUM: streams: Don't switch the SI to SI_ST_DIS if we have data to send.
In sess_established(), don't immediately switch the backend stream_interface
to SI_ST_DIS if we only got a SHUTR. We may still have something to send,
ie if the request is a POST, and we should be switched to SI_ST8DIS later
when the shutw will happen.

This should be backported to 2.0 and 1.9.

(cherry picked from commit 7859526fd6ce7ea33e20b7e532b21aa2465cb11d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>