MINOR: hlua: Don't set request analyzers on response channel for lua actions
Setting some requests analyzers on the response channel was an old trick to be
sure to re-evaluate the request's analyers after the response's ones have been
called. It is no more necessary. In fact, this trick was removed in the version
1.8 and backported up to the version 1.6.

This patch must be backported to all versions since 1.6 to ease the backports of
fixes on the lua code.

(cherry picked from commit 51fa358432247fe5d7259d9d8a0e08d49d429c73)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>