BUG/MAJOR: http/sample: use a static buffer for raw -> htx conversion
Multiple calls to smp_fetch_fhdr use the header context to keep track of
header parsing position; however, when using header sampling on a raw
connection, the raw buffer is converted into an HTX structure each time, and
this was done in the trash areas; so the block reference would be invalid on
subsequent calls.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit 458eafb36df88932a02d1ce7ca31832abf11b8b3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>