BUG/MEDIUM: proxy: Don't forget the SF_HTX flag when upgrading TCP=>H1+HTX.
In stream_end_backend(), if we're upgrading from TCP to H1/HTX, as we don't
destroy the stream, we have to add the SF_HTX flag on the stream, or bad
things will happen.
This was broken when attempting to fix github issue #196.

This should be backported to 2.0.

(cherry picked from commit 71b20c26bef68e2c7d52f1cfcdc7321d12c58e9e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>