BUG/MEDIUM: mux_h1: Don't bother subscribing in recv if we're not connected.
In h1_recv(), if we're still waiting for the connection, don't bother to
subscribe, just return 0.
We will be called back later because we will be subscribed for sending anyway,
and doing so may lead to the tasklet being waked in a loop until the
connection is established, as, by just subscribing without attempting to
receive, we will never leave the fd cache, and so we will always consider
we're ready for receiving.

This is a direct commit to the 2.0 branch, as the problem doesn't exist in
master. It shouldn't be backported.