BUG/MEDIUM: url32 does not take the path part into account in the returned hash.
The url32 sample fetch does not take the path part of the URL into
account. This is because in smp_fetch_url32() we erroneously modify
path.len and path.ptr before testing their value and building the
path based part of the hash.

This fixes issue #235

This must be backported as far as 1.9, when HTX was introduced.

(cherry picked from commit 2dd26ca9ff8e642611b8b012d6aee45ea45196bc)
[wt: adjusted context, we still have legacy in 2.0]
Signed-off-by: Willy Tarreau <w@1wt.eu>