BUG/MINOR: checks: stop polling for write when we have nothing left to send
Since the change of I/O direction, we perform the connect() call and
the send() call together from the top. But the send call must at least
disable polling for writes once it does not have anything left to send.

This bug is partially responsible for the waste of resources described
in issue #253.

This must be backported to 2.0.

(cherry picked from commit 5909380c057cef2f54a3cd3f4ff249649c6fe101)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>